+++ tests/protocols/gmpv208/entities/tasks/test_create_task.py
@@ -22,9 +22,6 @@ from gvm.errors import RequiredArgument, InvalidArgument, InvalidArgumentType
 
 from gvm.protocols.gmpv208 import HostsOrdering
 
-from ...gmpv208 import Gmpv208TestCase
-from ...gmpv214 import Gmpv214TestCase
-
 
 class GmpCreateTaskTestCase:
     def test_create_task(self):
@@ -408,12 +405,3 @@ class GmpCreateTaskTestCase:
                 scanner_id='s1',
                 observers='',
             )
-
-
-# For new versions add another Mixin here.
-class Gmpv208CreateTaskTestCase(GmpCreateTaskTestCase, Gmpv208TestCase):
-    pass
-
-
-class Gmpv214CreateTaskTestCase(GmpCreateTaskTestCase, Gmpv214TestCase):
-    pass
