+++ tests/protocols/gmpv208/entities/tasks/test_stop_task.py
@@ -18,28 +18,16 @@
 
 from gvm.errors import GvmError
 
-from ...gmpv208 import Gmpv208TestCase
-from ...gmpv214 import Gmpv214TestCase
 
+class GmpStopTaskTestCase:
+    def test_stop_task(self):
+        self.gmp.stop_task('a1')
 
-class GmpStartTaskTestCase:
-    def test_start_task(self):
-        self.gmp.start_task('a1')
-
-        self.connection.send.has_been_called_with('<start_task task_id="a1"/>')
+        self.connection.send.has_been_called_with('<stop_task task_id="a1"/>')
 
     def test_missing_id(self):
         with self.assertRaises(GvmError):
-            self.gmp.start_task(None)
+            self.gmp.stop_task(None)
 
         with self.assertRaises(GvmError):
-            self.gmp.start_task('')
-
-
-# For new versions add another Mixin here.
-class Gmpv208StartTaskTestCase(GmpStartTaskTestCase, Gmpv208TestCase):
-    pass
-
-
-class Gmpv214StartTaskTestCase(GmpStartTaskTestCase, Gmpv214TestCase):
-    pass
+            self.gmp.stop_task('')
